Re-factor.
