+++ app/controllers/concerns/spree/admin/public_metadata_concern.rb
@@ -19,6 +19,22 @@ module Spree
             end
           end
         end
+
+        private
+
+        def assert_public_metadata(object)
+          return unless params[:public_metadata].present?
+
+          params[:public_metadata][:key].each_with_index do |key, i|
+            next unless key.present?
+
+            object.public_metadata[format_key(key)] = params[:public_metadata][:value][i]
+          end
+        end
+
+        def format_key(key)
+          key.downcase.parameterize.underscore.to_sym
+        end
       end
     end
   end
